# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MYSIG API",
        "description": "MYSIG API for Frontend",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/claim/car-policy": {
            "post": {
                "tags": [
                    "Claim"
                ],
                "summary": "ClaimForCar",
                "description": "Show Car Claim API",
                "operationId": "ClaimsForCar",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "incident_date": {
                                        "type": "string"
                                    },
                                    "incident_location": {
                                        "type": "string"
                                    },
                                    "policy_id": {
                                        "type": "string"
                                    },
                                    "incident_description": {
                                        "type": "string"
                                    },
                                    "insurer_id": {
                                        "type": "string"
                                    },
                                    "insurer_policy_id": {
                                        "type": "string"
                                    },
                                    "policy_type": {
                                        "type": "string"
                                    },
                                    "image_1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image_2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image_3": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image_4": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/policy/mypolicy-document": {
            "post": {
                "tags": [
                    "Policy Document"
                ],
                "summary": "my policy document link",
                "description": "my policy document link",
                "operationId": "policydocument",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "company_details": {
                                        "type": "string"
                                    },
                                    "policy_name": {
                                        "type": "string"
                                    },
                                    "policy_type_name": {
                                        "type": "string"
                                    },
                                    "policy_number": {
                                        "type": "string"
                                    },
                                    "start_date": {
                                        "type": "string"
                                    },
                                    "end_date": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}